+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
-## 0.0.0 (2020-04-21)
+## 1.0.0 (2020-04-21)
 
 - initial project structure based on [geronimo-iia/template-python](https://github.com/geronimo-iia/template-python)
+- integrate old stuff about rx, async
+- use python 3.8 and protocol
+- add a bunch of test unit
+
+